# src/CHC.Consent/CHC.Consent.Api.Client/generated/Models/CompositeIdentifierType.cs
// <auto-generated>
// (C) 2018 CHC  License: TBC
// </auto-generated>

namespace CHC.Consent.Api.Client.Models
{
    using Newtonsoft.Json;
    using System.Collections;
    using System.Collections.Generic;
    using System.Linq;

    public partial class CompositeIdentifierType : IIdentifierType
    {
        /// <summary>
        /// Initializes a new instance of the CompositeIdentifierType class.
        /// </summary>
        public CompositeIdentifierType()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the CompositeIdentifierType class.
        /// </summary>
        public CompositeIdentifierType(string systemName = default(string), IList<IDefinition> identifiers = default(IList<IDefinition>))
            : base(systemName)
        {
            Identifiers = identifiers;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "identifiers")]
        public IList<IDefinition> Identifiers { get; private set; }

    }
}