# src/CHC.Consent/CHC.Consent.Api.Client/generated/Models/IDefinition.cs
// <auto-generated>
// (C) 2018 CHC  License: TBC
// </auto-generated>

namespace CHC.Consent.Api.Client.Models
{
    using Newtonsoft.Json;
    using System.Linq;

    public partial class IDefinition
    {
        /// <summary>
        /// Initializes a new instance of the IDefinition class.
        /// </summary>
        public IDefinition()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the IDefinition class.
        /// </summary>
        public IDefinition(string systemName = default(string), IIdentifierType type = default(IIdentifierType))
        {
            SystemName = systemName;
            Type = type;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "systemName")]
        public string SystemName { get; private set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "type")]
        public IIdentifierType Type { get; private set; }

    }
}